源文件/文件夹不存在时抛出错误

+++ zopen/frs/frs.py
@@ -298,6 +298,9 @@ class FRS(ArchiveSupportMixin, RecycleMixin, CacheMixin):
 
         don't keep stat
         """
+        if not self.exists(src):
+            raise ValueError("source path is not exists: %s" % src)
+            
         if self.isfile(src):
             self.copyfile(src, dst)
         else:
